feat(database): rebuilds offsets daily at midnight
every night the database offsets get thrown off. these functions change the offset immediately every
night at midnight, then begin rebuilding the database. this happens sequentially so that during the
rebuilding process requests can be properly handled with either a -1 offset for entries that have
not been rebuilt yet, a normal offset for entries that have been rebuilt, or a 200ms timeout for
entries that are about to be rebuilt.